huffman in a new file
